Move spinner to base components so it's available outside of the checkout package
