+++ assets/js/base/components/spinner/index.tsx
