+++ assets/js/base/components/spinner/style.scss
