+++ packages/checkout/index.js
@@ -8,5 +8,4 @@ export { default as ExperimentalOrderShippingPackages } from './order-shipping-p
 export { default as Panel } from './panel';
 export { SlotFillProvider } from 'wordpress-components';
 export { default as Button } from './button';
-export { default as Spinner } from './spinner';
 export { default as Label } from './label';
